Move the onboarding doc alongside the other events.
